Refactor hand parameter to hand_analyzer in winning hand checks
- Updated function signatures in check_2_han.rs, check_3_han.rs, check_5_han.rs, check_6_han.rs, check_yakuman.rs, and checker.rs to use `hand_analyzer` instead of `hand`.
- Adjusted internal logic to reference `hand_analyzer` for consistency across winning hand checks.

+++ src/hand_info/hand_analyzer.rs
@@ -544,7 +544,7 @@ fn count_normal_shanten_recursively(
         *sequential2_result = sequential2.clone();
         single_result.clear();
         for i in Tile::M1..=Tile::Z7 {
-            for _ in 0..summarized_hand[i as usize] {
+            for _ in 0..summarized_hand[i as usize] as u32 {
                 single_result.push(i);
             }
         }
